cast double (for e_zoom comp) + fix issue for color_parsing

+++ srcs/window/loop/events/e_zoom.c
@@ -6,7 +6,7 @@
 /*   By: rbroque <rbroque@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/16 17:41:52 by rbroque           #+#    #+#             */
-/*   Updated: 2023/01/23 12:28:59 by rbroque          ###   ########.fr       */
+/*   Updated: 2023/03/02 16:57:34 by rbroque          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,14 +19,14 @@ static void	change_zoom(t_win *window, long zoom_shift)
 
 int	increase_zoom(t_win *window)
 {
-	if (window->map.zoom < MAX_ZOOM - ZOOM_SHIFT)
+	if (window->map.zoom < (double)(MAX_ZOOM - ZOOM_SHIFT))
 		change_zoom(window, ZOOM_SHIFT);
 	return (EXIT_SUCCESS);
 }
 
 int	decrease_zoom(t_win *window)
 {
-	if (window->map.zoom > MIN_ZOOM + ZOOM_SHIFT)
+	if (window->map.zoom > (double)(MIN_ZOOM + ZOOM_SHIFT))
 		change_zoom(window, -ZOOM_SHIFT);
 	return (EXIT_SUCCESS);
 }
